requirements.txt
python-slugify[unidecode]
podcastparser
lxml